backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.2.0
httpx>=0.25.0

# Redis
redis>=5.0.0

# Vector DB
qdrant-client>=1.7.0

# AI/LLM
openai>=1.6.0

# Rate limiting
slowapi>=0.1.9

# Utilities
httpx>=0.25.0
aiohttp>=3.9.0
